requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
gunicorn==21.2.0
requests==2.31.0
dj-database-url==2.1.0
drf-yasg==1.21.7
pytz==2024.1
APScheduler==3.10.4
django-apscheduler==0.6.2

